+++ changelog.md
@@ -68,6 +68,7 @@ Codion Change Log
 - AbstractColumnConditionPanel and AbstractFilterTableConditionPanel added, related refactoring.
 - AbstractColumnConditionPanel, AbstractFilterTableConditionPanel.focusGainedEvent() removed.
 - AbstractColumnConditionPanel and AbstractFilterTableConditionPanel replaced with interfaces FilterTable.ColumnConditionPanel and FilterTable.TableConditionPanel, related refactoring.
+- TableConditionPanel.conditionPanel() now returns Optional.
 ### is.codion.framework.domain
 - DefaultDomain renamed DomainModel.
 - Entity.valuesEqual() instance methods renamed equalValues(), static ones removed.